[omnibox] Improve safety of ShortcutsDatabase int to enum casting
In the wild, there are definitely a few users (like dpapad@) with
corrupt Shortcuts sqlite3 tables. Probably Chrome crashed while the
file was being written to disk, or bit rot, or a cosmic ray.

This causes crashes, because the integral values of enums like
'type' might not correspond to an in-range AutocompleteMatchType.

Right now, we just static cast that number to the enum, but this could
lead to a crash later when the user tries to get the vector icon for
this match (as the enum value would be out of bounds).

This CL:

 1. Enhances the crash logging of AutocompleteMatch::GetVectorIcon to
    give us a more definitive picture of which types of matches and
    from what provider is causing the crashes.

 2. Updates the ShortcutDatabase API so that callers no longer see the
    raw integers coming out of SQL, but rather does the range-check
    and conversion right away.

    The values are "toxic waste" until validated, so we don't want
    any callers to have to deal with that complexity.

Bug: 1024114
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2667785
Commit-Queue: Tommy Li <tommycli@chromium.org>
Cr-Commit-Position: refs/heads/master@{#852295}

+++ components/omnibox/browser/shortcuts_backend.cc
@@ -178,8 +178,7 @@ ShortcutsDatabase::Shortcut::MatchCore ShortcutsBackend::MatchToMatchCore(
 
   return ShortcutsDatabase::Shortcut::MatchCore(
       normalized_match->fill_into_edit, normalized_match->destination_url,
-      static_cast<int>(normalized_match->document_type),
-      normalized_match->contents,
+      normalized_match->document_type, normalized_match->contents,
       StripMatchMarkers(normalized_match->contents_class), description,
       StripMatchMarkers(description_class), normalized_match->transition,
       match_type, normalized_match->keyword);
